Assignment Redo API
Reworked page to use getscoresfromusers API call. Now only uses one API call to run the page.
There is a delay since I am doing the call in-page now and not before, so added a basic shimmer effect that can be redone later.
Fixed and tested push notification code to work with the new page style.
Made some other minor adjustments to the code and page.