Bottom Padding Issue Fix
So, I got rid of the bottom padding on the login (and other collapsible scroll pages).
The collapsing app bar now uses a custom scroll view, except on the assignment page which broke it.
By using the appropriate sliver widgets it fixed the issue with padding, but the behavior may not be intended.