Webview Resizing Bugfix
I modified the viewport to use the width minus a margin instead of a static 400. In addition, I added a slight delay to the injectviewport call that way it'd give time for the javascript reader to work. This was less delay than adding it to the onLoadStop, and only runs once as opposed to haivng in run in both functions which had a similiar result.