Question Url Navigation
Navigation to questions using URLs and cookie, as opposed to the iframe-based solution.

Cookie is created on login, and stored. Cookie is used on original opening of the question to create a jwt, then it renavigates using a redirect. After the first navigation, it navigates using URLs using the jwt without needing to reauthenticate in the webview.