Fix App Bar Position
The actual fix was simply making sure center title was set to off. However, the margins in IOS for the title were messed up.

I played around quite a bit to avoid doing an IOS-specific solution. IOS was about 20 pixels off on the constraints given, possibly due to a safe area or something, thus just made top 20 lower when setting it, and added a min check in getDiff to keep the margin at the right spot for the title