Safe area
Added a safe area and some padding to the bottom. For some reason, even with a safe area,  IOS was still slightly off in spacing.

I added a variable to make the UI consistent across our testing devices, but I understand that's not the best method and we could replace it with a solid 8, the android devices would just be slightly higher from the bottom.