Notifications 1
Sort of fixed notifcations, hard to test though.

Notifications that show up in background are now saved by the app and will show up in notifications window.

When Selected from notification window, it opens to the specified page, and deletes the selected one.

Switched to using a Hive database as opposed to preferences to store notifications.